Revert "Add data loader for field"
